Error while building the NonGNU ELPA tarball of org-auto-tangle
Hi,

As the `Subject:` suggests, elpa.nongnu.org encountered an error while
building the tarball, because `org-auto-tangle.el` fails to include
a valid email address to send notifications.

The patch below fixes that; it also comes with an unrelated and untested
patch for `org-auto-tangle-async` which makes it use closures rather than
quoted lambdas (prompted by a parenthesis error in the previous code,
and then I got carried away :-).

        Stefan