Pass user-specified variables to the async process
Augment the `auto_tangle' keyword to handle a 'vars' option which
contains a colon-separated list of variable names that will be passed
and preserved in the async tangling process.

Add an `org-auto-tangle' configuration group.

Add an `org-auto-tangle-with-vars' configuration option.

Support disabling auto-tangling by adding "nil" to AUTO_TANGLE keyword
without having to remove any variables that are set there.